refactor(ui): align dashboard cards with AppShell design system
