+++ src/components/common/SearchableTickerDropdown.css
@@ -1,0 +1,75 @@
+.chart-search-dropdown {
+  position: relative;
+}
+
+.chart-select {
+  border-radius: 999px;
+  border: 1px solid rgba(148, 163, 184, 0.9);
+  padding: 4px 10px;
+  font-size: 12px;
+  background: #ffffff;
+  color: #111827;
+  cursor: pointer;
+}
+
+.chart-select--button {
+  display: inline-flex;
+  align-items: center;
+  gap: 4px;
+}
+
+.chart-select-caret {
+  font-size: 10px;
+  opacity: 0.6;
+}
+
+.chart-select-menu {
+  margin-top: 4px;
+  border: 1px solid rgba(148, 163, 184, 0.4);
+  background: white;
+  border-radius: 8px;
+  padding: 6px;
+  position: absolute;
+  z-index: 1000;
+  width: 180px;
+  box-shadow: 0 2px 6px rgba(0,0,0,0.08);
+}
+
+.chart-search-input {
+  width: 100%;
+  border: 1px solid #d1d5db;
+  padding: 4px 6px;
+  border-radius: 6px;
+  margin-bottom: 6px;
+  font-size: 12px;
+}
+
+.chart-select-options {
+  list-style: none;
+  margin: 0;
+  padding: 0;
+  max-height: 220px;
+  overflow-y: auto;
+}
+
+.chart-select-option {
+  padding: 4px 6px;
+  border-radius: 4px;
+  font-size: 12px;
+  cursor: pointer;
+}
+
+.chart-select-option:hover {
+  background: #f0f4f8;
+}
+
+.chart-select-option--active {
+  background: #e0f2fe;
+  font-weight: 600;
+}
+
+.chart-select-option--empty {
+  padding: 6px;
+  color: #6b7280;
+  text-align: center;
+}
